+++ tests-integration/src/lib.rs
@@ -1,4 +1,4 @@
 use doc_comment::doc_comment;
 
-// #[doc = include_str!("../../README.md")]
+#[cfg(feature = "full")]
 doc_comment!(include_str!("../../README.md"));
